TP-5487: backport to working gnupg tree

+++ install.sh
@@ -8,7 +8,7 @@ set -o xtrace
 sed -i '/deb-src/s/^# //' /etc/apt/sources.list
 apt update
 
-apt install -y dpkg-dev
+apt install -y dpkg-dev sudo
 DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends tzdata
 
 mkdir -p ~/sources
@@ -31,7 +31,14 @@ apt install -y \
     gsfonts \
     transfig \
     texinfo \
+    ssh \
     openssh-server
+
+apt-get update -q -q
+apt-get install --yes rsync build-essential git gpg automake autoconf gettext libtool \
+    libz-dev libbz2-dev libldap2-dev libsqlite3-dev libgnutls28-dev libcurl4-gnutls-dev libreadline-dev librsvg2-bin libusb-1.0-0-dev \
+    texinfo transfig fig2dev imagemagick file ghostscript swig doxygen graphviz \
+    pkg-config autopoint python-all-dev python3-all-dev qtbase5-dev
 
 # cat ~/.config/ImageMagick/policy.xml 
 # <policymap>
@@ -50,9 +57,10 @@ apt install -y \
 # sudo systemctl status ssh
 
 
-mkdir -p /usr/local/bin/
-ln /usr/bin/pinentry /usr/local/bin/pinentry 
-
+if [ ! -f /usr/local/bin/pinentry ]; then
+    mkdir -p /usr/local/bin/
+    ln /usr/bin/pinentry /usr/local/bin/pinentry 
+fi
 
 echo 0 | sudo tee /proc/sys/kernel/yama/ptrace_scope
